Split ksGpuDevice_Create into physical/logical
Add a control bitmask to the ksGpuDevice_Create functiont that allows a
caller to either create the physical device, the logical device, or
both.

This is a prerequisite for display surface functionality, as the logical
device implicitly creates the swapchain, but the physical device is
required to query the surface creation parameters.