Added support for VkDisplaySurfaceKHR.
This change adds a new "platform", OS_NEUTRAL_DISPLAY_SURFACE. Currently
we don't support input in this mode, as input is os-specific.

I have a larger change planned that separates the input from
presentation, and further, allows runtime-selection of presentation
(e.g.  allow the sample to be built for both xlib and xcb, choosing
which path to use via a command line argument).